Update diablo.go
Tried to make Chaos Sanctuary runs a bit safer when popping the seals. It moves to specific safe coordinates after popping a seal to spawn the elite, based on the location of the seal coordinates.

Added extra forced buff times to make sure we always have BO.

Sometimes the bot wouldn't go back to the star to kill Diablo, added 3 extra coordinates to use to walk back to the star.