Fix Seal interactions (chaos) Prevent bot from teleporting up to 5 time on seals before interacting.
change to ClearThroughpath : Increasing step.MoveTo(dest, step.WithDistanceToFinish(7) as option prevent not being to able to finish movement if our destination is center of a large object like Seal in diablo run. is used only for pathing, attack.go or pathfinder will use default DistanceToFinishMoving (4) .

Change to interaction.go :
Is similar to solution from ClearThroughPath :
if InteractObject with a Seal,  we increase DistanceToFinish so it will consider in range at this distance.  With default 4 it wont be able to reach destination and bounce back on seal since its further then 4 and we dont pass to step.InteractObject until we are in range.