Fixed loop that could get stuck
Removed ErrItemTooFar from not counting toward retries. There could be a loop where the item is too far away and the logic does not move enough or try other moving techniques to get closer. This allows other pickup attempts using moving logic and a movebeyond when the item is more then 7 away.