New Hover Lookup, Increase Sprial Attempts
Changed the way item pickup checks for hover status. Previously refreshgamedata was used to add hover status to items. Now checking hover status directly and checking if hovered unitid matches the item unitid we want to pickup. This greatly increases the speed of item pickup actions, thus an increase to spiral attempts can be added without the extra overhead of constantly refreshing all game data and waiting for it to return. Removed an unused spiral delay.